src/requirements.txt
Django==4.0.5
django-cors-headers==3.13.0
django-webpack-loader==1.5.0
djangorestframework==3.13.1
gunicorn==20.1.0
psycopg2
requests==2.28.0
sqlparse==0.4.2
asgiref==3.5.0
django-phonenumber-field==7.0.0
django-phonenumbers==1.0.1
phonenumbers==8.12.55
pycparser==2.21
pytz==2022.2.1
sqlparse==0.4.2
cffi==1.15.1
cryptography==38.0.1
django-rest-knox==4.2.0
djangorestframework-simplejwt
